refactor : delete files
프로젝트와 상관없는 파일 삭제
