--- server/src/main/java/toyspringboot/server/Domain/Entity/Messages.java
@@ -1,19 +1,0 @@
-package toyspringboot.server.Domain.Entity;
-
-import lombok.*;
-import javax.persistence.*;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class Messages {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String message;
-}
