--- server/src/main/java/toyspringboot/server/Domain/Repository/MessageRepository.java
@@ -1,10 +1,0 @@
-package toyspringboot.server.Domain.Repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import toyspringboot.server.Domain.Entity.Messages;
-
-@Repository
-public interface MessageRepository extends JpaRepository<Messages, Long> {
-
-}
