+++ server/src/main/java/toyspringboot/server/Swagger/SwaggerConfig.java
@@ -1,8 +1,7 @@
-package toyspringboot.server.Config;
+package toyspringboot.server.Swagger;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
@@ -10,9 +9,6 @@ import springfox.documentation.service.ApiInfo;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import java.util.HashSet;
-import java.util.Set;
 
 @Configuration
 @EnableSwagger2
